Project for examining diffusion and heat equation.
This project uses python to model the heat equation and diffusion equation. We closely examine how the forward-euler method compares against the analytical solutions for these equations.  We assess how the error behaves based on step size and consider how to interpret these graphs in a biological setting. Plotting was all done in matplotlib.