Modeling the growth of an animal species.
In this project we model the spread of a starling population using the 1-d advection equation. This models the growth of a population over a distance as a wave form where the height of the wave corresponds to the carrying capacity of the animal species at that distance. We examine how the boundary conditions effect the efficiency of the forward and backward differencing methods. Issues with the differencing methods are demonstrated including a total breakdown of the model in the forward difference and a failure to capture the smoothness of the true wave equation.